+++ metadata.rb
@@ -1,4 +1,4 @@
-# Copyright 2015 Joshua Timberman
+# Copyright 2015 Joshua Timterman
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -16,9 +16,9 @@ name             'vagrant'
 maintainer       'Sous Chefs'
 maintainer_email 'help@sous-chefs.org'
 license          'Apache-2.0'
-description      'Installs Vagrant and provides a vagrant_plugin LWRP for installing Vagrant plugins.'
+description      'Installs Vagrant and provides a vagrant_plugin resource for installing Vagrant plugins.'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.9.0'
+version          '0.9.1'
 
 source_url       'https://github.com/sous-chefs/vagrant'
 issues_url       'https://github.com/sous-chefs/vagrant/issues'
@@ -31,6 +31,5 @@ supports         'windows'
 supports         'mac_os_x'
 
 depends          'dmg', '>= 2.2.2'
-depends          'windows', '>= 1.38'
 
 chef_version     '>= 13.0'