+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Vagrant Cookbook Changelog
 
+## 0.9.1
+* Update the README to describe the new vagrant resource
+
 ## 0.9.0
 * Create a custom resource to install vagrant.  Fix issue #69